src/asociaciones_estrategicas/asociaciones-requirements.txt
attrs==25.3.0
click<8.2
coverage==7.9.2
flask[async]==3.1.1
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
greenlet<3.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
PyMySQL==1.1.1
PyMySQL[rsa]==1.1.1
pluggy==1.6.0
pulsar-client==3.7.0
pulsar-client[avro]==3.7.0
py==1.11.0
PyDispatcher==2.0.7
pyparsing==3.2.3
pytest==8.4.1
PyYAML==6.0.2
SQLAlchemy==2.0.41
tomli==2.2.1
Werkzeug==3.1.3